app/daawat/service/firebase_service.py
import pyrebase

config = {
    "apiKey": "AIzaSyCi5-vfmk7VW2xFtgacj3_arjL04KeaW2A",
    "authDomain": "dawaat-9e0f1.firebaseapp.com",
    "projectId": "dawaat-9e0f1",
#     "databaseURL":"https://dawaat-9e0f1-default-rtdb.firebaseio.com",
    "storageBucket": "dawaat-9e0f1.appspot.com",
    "messagingSenderId": "55134356040",
    "appId": "1:55134356040:web:ea4f51040695d9d59f0e7d",
    "measurementId": "G-Z2N8MPTMYK"
};

firebase = pyrebase.initialize_app(config)
storage = firebase.storage()
