app/requirements.txt
asgiref==3.3.1
cassandra-driver==3.24.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.1.4
django-cassandra-db==0.0.2
gcloud==0.18.3
geomet==0.2.1.post1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jws==0.1.3
oauth2client==4.1.3
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.9
pypng==0.0.20
PyQRCode==1.2.1
Pyrebase4==4.4.3
python-jwt==2.0.1
pytz==2020.5
qrcode==6.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
